SolicitudService/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
tenacity==8.2.3